Update stackdriver to new form, introduce the skeleton of logging, and a facade over the two
