+++ adapter/stackdriver/config/BUILD
@@ -22,7 +22,7 @@ gogoslick_proto_library(
         "config.proto",
     ],
     verbose = 0,
-    visibility = ["//adapter/stackdriver:__pkg__"],
+    visibility = ["//adapter/stackdriver:__subpackages__"],
     deps = [
         "@com_github_gogo_protobuf//gogoproto:go_default_library",
         "@com_github_gogo_protobuf//sortkeys:go_default_library",
