+++ adapter/stackdriver/log/BUILD
@@ -1,0 +1,11 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["log.go"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/adapter:go_default_library",
+        "//template/logentry:go_default_library",
+    ],
+)
